UI and Bug Fixes
- Chnaged scroll behavious from bottom to top
- Resolved Delete Chat bug (by creating new route inside /api/chat)
- Removed Substring from mail subject and used css property for text cliping
- Increased the intensity of the background color of skeleton
