+++ components/custom/document-skeleton.tsx
@@ -3,13 +3,13 @@
 export const DocumentSkeleton = () => {
   return (
     <div className="flex flex-col gap-4 w-full">
-      <div className="animate-pulse rounded-lg h-12 bg-muted-foreground/20 w-1/2" />
-      <div className="animate-pulse rounded-lg h-5 bg-muted-foreground/20 w-full" />
-      <div className="animate-pulse rounded-lg h-5 bg-muted-foreground/20 w-full" />
-      <div className="animate-pulse rounded-lg h-5 bg-muted-foreground/20 w-1/3" />
+      <div className="animate-pulse rounded-lg h-12 bg-muted-foreground/40 w-1/2" />
+      <div className="animate-pulse rounded-lg h-5 bg-muted-foreground/40 w-full" />
+      <div className="animate-pulse rounded-lg h-5 bg-muted-foreground/40 w-full" />
+      <div className="animate-pulse rounded-lg h-5 bg-muted-foreground/40 w-1/3" />
       <div className="animate-pulse rounded-lg h-5 bg-transparent w-52" />
-      <div className="animate-pulse rounded-lg h-8 bg-muted-foreground/20 w-52" />
-      <div className="animate-pulse rounded-lg h-5 bg-muted-foreground/20 w-2/3" />
+      <div className="animate-pulse rounded-lg h-8 bg-muted-foreground/40 w-52" />
+      <div className="animate-pulse rounded-lg h-5 bg-muted-foreground/40 w-2/3" />
     </div>
   );
 };
