+++ components/custom/sidebar-history.tsx
@@ -105,7 +105,7 @@ export function SidebarHistory({ user }: { user: User | undefined }) {
   const [showDeleteDialog, setShowDeleteDialog] = useState(false);
   const router = useRouter();
   const handleDelete = async () => {
-    const deletePromise = fetch(`/api/chat?id=${deleteId}`, {
+    const deletePromise = fetch(`/api/chat/delete?id=${deleteId}`, {
       method: 'DELETE',
     });
